audit: fix correctness bugs
Fix six bugs found during code audit:

- world.h: fix typo in fallback #define (PHASE2_BAKCEND →
  PHASE2_BACKEND).  The misspelling caused the guard to
  define the wrong symbol; the code worked only because
  undefined macros evaluate to 0 in #if directives.

- world_cuda.c: fix type mismatch in
  world_backend_destroy() parameter (struct world →
  struct world_info).  The nonexistent struct world broke
  CUDA backend compilation.

- data.c: cast unsigned char to uint64_t before left shift
  in data_multidet_foreach().  Without the cast, the byte
  value is promoted to int and shifting by j >= 31 is
  undefined behaviour.

- circ.c, qdrift.c, cmpsit.c: uncomment cleanup calls in
  error-unwind paths of circ_init(), qdrift_init(), and
  cmpsit_init()/ranct_init().  The commented-out frees
  caused memory leaks on partial initialisation failure.

- b-qreg.c: move q.i and q.amp assignment before the
  qreg_setamp() call that uses them.  The previous order
  passed uninitialised values.

- ph2run.c: fix printf format string in cmd_qdrift_write().
  step_size (double) was printed with %zu and depth
  (size_t) with %.3f due to swapped format/argument order.

+++ phase2/data.c
@@ -248,7 +248,7 @@ int data_multidet_foreach(data_id fid,
 	for (size_t i = 0; i < ndets; i++) {
 		uint64_t idx = 0;
 		for (size_t j = 0; j < nqb; j++) {
-			idx += dets[i * nqb + j] << j;
+			idx += (uint64_t)dets[i * nqb + j] << j;
 		}
 		_Complex double cf = CMPLX(cfs[2 * i], cfs[2 * i + 1]);
 		rc = op(cf, idx, op_data);
